+++ src/config/keybinds.rs
@@ -333,6 +333,9 @@ pub struct Keybinds {
     pub split_horizontal: ActionKeybinds,
     pub close_pane: ActionKeybinds,
     pub zoom: ActionKeybinds,
+    pub stack_pane: ActionKeybinds,
+    pub unstack_pane: ActionKeybinds,
+    pub break_pane: ActionKeybinds,
     pub resize_mode: ActionKeybinds,
     pub toggle_sidebar: ActionKeybinds,
     pub custom_commands: Vec<CustomCommandKeybind>,
@@ -495,6 +498,9 @@ impl Config {
             split_horizontal: empty_action!(),
             close_pane: empty_action!(),
             zoom: empty_action!(),
+            stack_pane: empty_action!(),
+            unstack_pane: empty_action!(),
+            break_pane: empty_action!(),
             resize_mode: empty_action!(),
             toggle_sidebar: empty_action!(),
             custom_commands: Vec::new(),
@@ -636,6 +642,9 @@ impl Config {
             apply_action!(keybinds.split_horizontal, split_horizontal, source);
             apply_action!(keybinds.close_pane, close_pane, source);
             apply_action!(keybinds.zoom, zoom, source);
+            apply_action!(keybinds.stack_pane, stack_pane, source);
+            apply_action!(keybinds.unstack_pane, unstack_pane, source);
+            apply_action!(keybinds.break_pane, break_pane, source);
             apply_action!(keybinds.resize_mode, resize_mode, source);
             apply_action!(keybinds.toggle_sidebar, toggle_sidebar, source);